Multithreading (#7)
* Start working on multithreading for rendering part

* Implement correct (more or less) multithreading to start games together with application
Fully implement RenderTargetPanel control
Fix saving of gif images
Improve DrawingContext.cs
Extend EntityService.cs

* Make a lot improvements to image processing
Improve Measure and Arrange for Grid control
Rework loading of bitmap images
Make possibility for multiThreaded rendering for Game class inside UIApplication class
Made some improvements to AUML parser

* Rename image file types

* Improvements to IRawBitmap tto support saving to file

* Changes in Loading/Saving images
Improvements to Grid control measure/arrange logic
Improvements to rendering performance

* Improvements to DrawingContext class to use hash instead of direct comparing of input parameters

+++ Adamantium/Adamantium.Engine.Graphics/Effects/EffectPool.cs
@@ -394,7 +394,7 @@ namespace Adamantium.Engine.Graphics.Effects
 
         private static Buffer DefaultConstantBufferAllocator(GraphicsDevice device, EffectPool pool, EffectConstantBuffer constantBuffer)
         {
-            return Buffer.Uniform.New(device, constantBuffer.BackingBuffer.Size);
+            return Buffer.Uniform.New(device, (int)constantBuffer.BackingBuffer.Size);
         }
 
         private void OnEffectAdded(EffectPoolEventArgs e)
